+++ ui/src/app/edge/live/common/autarchy/autarchy.component.html
