+++ ui/src/app/edge/live/flat/flat-widget-line/flat-widget-percentagebar.html
@@ -4,6 +4,6 @@
         <rect *ngIf="displayValue" x="1" y="2" rx="5" ry="5" attr.width="{{displayValue}}%" height="16"
             style="fill:#2d8fab" />
         <text x="50%" y="58%" dominant-baseline="middle" text-anchor="middle" style="font-weight: 500">{{ displayValue |
-            unitvalue:'%' }}</text>
+            number:'1.0-0' }}&nbsp;%</text>
     </svg>
 </ng-container>
