+++ ui/src/app/edge/live/flat/flat-widget-line/flat-widget-percentagebar.ts
