+++ ui/src/app/edge/live/live.module.ts
@@ -1,13 +1,12 @@
-import { AdministrationComponent } from './evcs/modal/administration/administration.component';
-import { AsymmetricPeakshavingComponent } from './peakshaving/asymmetric/asymmetricpeakshaving.component';
-import { AsymmetricPeakshavingModalComponent } from './peakshaving/asymmetric/modal/modal.component';
-import { AutarchyComponent } from './autarchy/autarchy.component';
-import { AutarchyModalComponent } from './autarchy/modal/modal.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { SharedModule } from './../../shared/shared.module';
 import { ChannelthresholdComponent } from './channelthreshold/channelthreshold.component';
 import { ChpSocComponent } from './chpsoc/chpsoc.component';
 import { ChpsocModalComponent } from './chpsoc/modal/modal.component';
+import { AutarchyComponent } from './common/autarchy/autarchy.component';
+import { AutarchyModalComponent } from './common/autarchy/modal/modal.component';
 import { ConsumptionComponent } from './consumption/consumption.component';
 import { ConsumptionModalComponent } from './consumption/modal/modal.component';
 import { Controller_Ess_FixActivePower } from './Controller_Ess_FixActivePower/Controller_Ess_FixActivePower';
@@ -19,14 +18,16 @@ import { DelayedSellToGridModalComponent } from './delayedselltogrid/modal/modal
 import { DigitalInputComponent } from './digitalinput/digitalinput.component';
 import { DigitalInputModalComponent } from './digitalinput/modal/modal.component';
 import { EnergymonitorModule } from './energymonitor/energymonitor.module';
-import { EvcsChart } from './evcsCluster/modal/evcs-chart/evcs.chart';
-import { EvcsClusterComponent } from './evcsCluster/evcsCluster.component';
 import { EvcsComponent } from './evcs/evcs.component';
+import { AdministrationComponent } from './evcs/modal/administration/administration.component';
 import { EvcsModalComponent } from './evcs/modal/modal.page';
 import { EvcsPopoverComponent } from './evcs/modal/popover/popover.page';
-import { FlatWidgetComponent } from './flat/flat-widget.component';
+import { EvcsClusterComponent } from './evcsCluster/evcsCluster.component';
+import { EvcsChart } from './evcsCluster/modal/evcs-chart/evcs.chart';
+import { ModalComponentEvcsCluster } from './evcsCluster/modal/evcsCluster-modal.page';
 import { FlatWidgetLine } from './flat/flat-widget-line/flat-widget-line';
-import { FlatWidgetPercentagebar } from './flat/flat-widget-percentagebar/flat-widget-percentagebar';
+import { FlatWidgetPercentagebar } from './flat/flat-widget-line/flat-widget-percentagebar';
+import { FlatWidgetComponent } from './flat/flat-widget.component';
 import { GridComponent } from './grid/grid.component';
 import { GridModalComponent } from './grid/modal/modal.component';
 import { HeatingElementComponent } from './heatingelement/heatingelement.component';
@@ -35,22 +36,21 @@ import { HeatPumpComponent } from './heatpump/heatpump.component';
 import { HeatPumpModalComponent } from './heatpump/modal/modal.component';
 import { InfoComponent } from './info/info.component';
 import { LiveComponent } from './live.component';
-import { ModalComponentEvcsCluster } from './evcsCluster/modal/evcsCluster-modal.page';
-import { NgModule } from '@angular/core';
 import { OfflineComponent } from './offline/offline.component';
-import { ProductionComponent } from './production/production.component';
-import { ProductionModalComponent } from './production/modal/modal.component';
-import { SelfConsumptionComponent } from './selfconsumption/selfconsumption.component';
-import { SelfconsumptionModalComponent } from './selfconsumption/modal/modal.component';
-import { SharedModule } from './../../shared/shared.module';
-import { SinglethresholdComponent } from './singlethreshold/singlethreshold.component';
-import { SinglethresholdModalComponent } from './singlethreshold/modal/modal.component';
-import { StorageComponent } from './storage/storage.component';
-import { StorageModalComponent } from './storage/modal/modal.component';
-import { SymmetricPeakshavingComponent } from './peakshaving/symmetric/symmetricpeakshaving.component';
+import { AsymmetricPeakshavingComponent } from './peakshaving/asymmetric/asymmetricpeakshaving.component';
+import { AsymmetricPeakshavingModalComponent } from './peakshaving/asymmetric/modal/modal.component';
 import { SymmetricPeakshavingModalComponent } from './peakshaving/symmetric/modal/modal.component';
-import { TimeslotPeakshavingComponent } from './peakshaving/timeslot/timeslotpeakshaving.component';
+import { SymmetricPeakshavingComponent } from './peakshaving/symmetric/symmetricpeakshaving.component';
 import { TimeslotPeakshavingModalComponent } from './peakshaving/timeslot/modal/modal.component';
+import { TimeslotPeakshavingComponent } from './peakshaving/timeslot/timeslotpeakshaving.component';
+import { ProductionModalComponent } from './production/modal/modal.component';
+import { ProductionComponent } from './production/production.component';
+import { SelfconsumptionModalComponent } from './selfconsumption/modal/modal.component';
+import { SelfConsumptionComponent } from './selfconsumption/selfconsumption.component';
+import { SinglethresholdModalComponent } from './singlethreshold/modal/modal.component';
+import { SinglethresholdComponent } from './singlethreshold/singlethreshold.component';
+import { StorageModalComponent } from './storage/modal/modal.component';
+import { StorageComponent } from './storage/storage.component';
 
 @NgModule({
   imports: [
